requirements-multi-agent.txt
# Azure AI Foundry dependencies for multi-agent support
azure-ai-agents>=1.0.0
azure-identity>=1.15.0

# Include base requirements
-r requirements.txt 